+++ src/apis/postAdminFitness/postAdminFitnessData.ts
@@ -1,0 +1,16 @@
+import { axiosInstance } from "../axios-instance";
+import config from "../config";
+
+export const postAdminFitnessData = async (formData: FormData) => {
+  const response = await axiosInstance.post(
+    `${config.apiBaseUrl}/admin/fitness/createFitness`,
+    formData,
+    {
+      headers: {
+        "Content-Type": "multipart/form-data",
+      },
+    }
+  );
+
+  return response.data;
+};
