+++ src/hooks/useAdminFitnessUpload.ts
@@ -1,5 +1,5 @@
 import { useMutation } from "@tanstack/react-query";
-import { postAdminFitnessData } from "../axios/postAdminFitnessData";
+import { postAdminFitnessData } from "../apis/postAdminFitness/postAdminFitnessData";
 
 export const useAdminFitnessUpload = () => {
   return useMutation({
